Split i18n, Item, Weapons and Profile into their own modules
