+++ profile.py
@@ -1,0 +1,3 @@
+class Profile:
+    """Represent global profile data"""
+    pass
